Fix Firebase Google Auth CSP, switch to redirect, and configure App Check debug token
