+++ .jules/bolt.md
@@ -4,3 +4,4 @@
 
 ## 2024-04-28 - [Интеграция Firebase Presence RTDB] **Урок:** При интеграции RTDB listener'ов нужно быть предельно внимательным к импортам (`import { rtdb } from '$lib/firebase'`) и обязательной отписке (`off()`) как от данных статуса, так и от `.info/connected` узла, чтобы избежать утечек памяти и исчерпания коннектов. **Действие:** Добавлена очистка всех подписок (`connectedUnsubscribe` и `presenceUnsubscribe`), а также проверено, что импорт базы берется именно из `$lib/firebase` как требует архитектура проекта.
 ## 2024-05-24 - [Fix Google Auth Popup & 404 Audio] **Урок:** Использование `signInWithPopup` может блокироваться браузерами после асинхронных операций (как Turnstile капча). Иногда лучшее решение — не менять рабочий флоу на `signInWithRedirect`, а перехватывать ошибку `auth/popup-blocked` и давать пользователю четкую инструкцию, как разрешить всплывающие окна в адресной строке. **Действие:** Добавил подробные UI-уведомления для заблокированных попапов и удалил обращение к отсутствующему файлу `entercasino.mp3`.
+## 2024-04-28 - [Fix Google Auth CSP & Redirect] **Урок:** При настройке CSP в SvelteKit (svelte.config.js) ключевые слова ('self', 'unsafe-inline') должны передаваться в двойных кавычках (например, "'self'"). Для обхода блокировки попапов Google Auth после асинхронной капчи надежнее использовать signInWithRedirect и getRedirectResult в глобальном макете. При этом важно корректно восстанавливать стейт сессии (userStore) и вызывать бекенд-синхронизацию токена. **Действие:** Настроил CSP, исправил флоу Google Auth с Popup на Redirect, перенес создание профиля в +layout.svelte с учетом синхронизации стора, добавил поддержку отладочного токена App Check через VITE-переменную.
