+++ src/routes/+layout.svelte
@@ -25,6 +25,12 @@
     import '$lib/i18n';
     import { waitLocale } from 'svelte-i18n';
 
+    import { auth, db } from '$lib/firebase';
+    import { getRedirectResult } from 'firebase/auth';
+    import { doc, getDoc, setDoc, serverTimestamp } from 'firebase/firestore';
+    import { getFunctions, httpsCallable } from 'firebase/functions';
+    import { userStore } from '$lib/stores';
+
     let themeState = 'default';
     let sideTextLeft = 'СТАТУС СИСТЕМЫ: ОНЛАЙН';
     let sideTextRight = 'МОЩНОСТЬ СЕТИ: 99%';
@@ -74,10 +80,128 @@
         }
     }
 
+    async function isUsernameAvailable(name: string): Promise<boolean> {
+        const trimmedName = name.trim();
+        if (trimmedName.length < 4) return false;
+        try {
+            const functions = getFunctions();
+            const checkUsernameFunc = httpsCallable(functions, 'checkUsername');
+            const result = await checkUsernameFunc({ username: trimmedName });
+            return (result.data as { isAvailable: boolean }).isAvailable;
+        } catch (e) {
+            console.error("Ошибка проверки username:", e);
+            return false;
+        }
+    }
+
     onMount(async () => {
         injectAnalytics({ mode: dev ? 'development' : 'production' });
         AudioManager.initialize();
         await waitLocale();
+
+        // Обработка Google Auth редиректа глобально
+        if (browser && auth) {
+            try {
+                const result = await getRedirectResult(auth);
+                if (result && result.user) {
+                    const user = result.user;
+                    console.log("✅ Google Auth Redirect Success:", user.uid);
+                    await user.getIdToken(true);
+
+                    const userDocRef = doc(db, "users", user.uid);
+                    let userDocSnap = await getDoc(userDocRef);
+
+                    if (!userDocSnap.exists()) {
+                        console.log("📝 Новый Google юзер (redirect), создаём профиль...");
+                        let generatedUsername = user.displayName || '';
+                        generatedUsername = generatedUsername.replace(/[^a-zA-Z0-9_]/g, '');
+                        if (generatedUsername.length < 3) generatedUsername = `user_${user.uid.substring(0, 8)}`;
+                        if (generatedUsername.length > 20) generatedUsername = generatedUsername.substring(0, 20);
+
+                        const isAvailable = await isUsernameAvailable(generatedUsername);
+                        if (!isAvailable) {
+                            const randomSuffix = Math.floor(Math.random() * 9999);
+                            generatedUsername = `${generatedUsername.substring(0, 15)}_${randomSuffix}`;
+                        }
+
+                        await setDoc(userDocRef, {
+                            username: generatedUsername,
+                            email: user.email || "",
+                            avatar_url: user.photoURL || "",
+                            about_me: "",
+                            social_link: "",
+                            createdAt: serverTimestamp(),
+                            casino_credits: 100,
+                            glitch_shards: 0,
+                            last_daily_bonus: null,
+                            owned_items: [],
+                            daily_streak: 0,
+                            isBanned: false,
+                            emailVerified: user.emailVerified,
+                            turnstileVerified: true
+                        });
+                        console.log("✅ Профиль Google создан!");
+
+                        const profileData = {
+                            uid: user.uid,
+                            username: generatedUsername,
+                            email: user.email || "",
+                            emailVerified: user.emailVerified,
+                            avatar_url: user.photoURL || "",
+                            social_link: "",
+                            about_me: "",
+                            status: "",
+                            casino_credits: 100,
+                            last_daily_bonus: null,
+                            daily_streak: 0,
+                            owned_items: [],
+                            equipped_frame: null,
+                            equipped_badge: null,
+                            equipped_bg: null,
+                            blocked_uids: []
+                        };
+                        userStore.set({ user: profileData, loading: false });
+                        await new Promise(resolve => setTimeout(resolve, 300));
+                    } else {
+                        console.log("♻️ Профиль Google уже существует");
+                        const data = userDocSnap.data();
+                        const profileData = {
+                            uid: user.uid,
+                            username: data.username,
+                            email: user.email || "",
+                            emailVerified: user.emailVerified,
+                            avatar_url: data.avatar_url || "",
+                            social_link: data.social_link || "",
+                            about_me: data.about_me || "",
+                            status: data.status || "",
+                            casino_credits: data.casino_credits ?? 100,
+                            last_daily_bonus: data.last_daily_bonus ? data.last_daily_bonus.toDate() : null,
+                            daily_streak: data.daily_streak || 0,
+                            owned_items: data.owned_items || [],
+                            equipped_frame: data.equipped_frame || null,
+                            equipped_badge: data.equipped_badge || null,
+                            equipped_bg: data.equipped_bg || null,
+                            blocked_uids: data.blocked_uids || []
+                        };
+                        userStore.set({ user: profileData, loading: false });
+                    }
+
+                    const token = await user.getIdToken();
+                    await fetch('/api/auth', {
+                        method: 'POST',
+                        headers: { 'Content-Type': 'application/json' },
+                        body: JSON.stringify({ idToken: token }),
+                    });
+                    await new Promise(resolve => setTimeout(resolve, 300));
+
+                    if ($page.url.pathname === '/login' || $page.url.pathname === '/register') {
+                        goto('/');
+                    }
+                }
+            } catch (e) {
+                console.error("❌ Ошибка при getRedirectResult:", e);
+            }
+        }
 
         const savedSeasonal = localStorage.getItem('protomap_seasonal_enabled');
         const isSeasonalEnabled = savedSeasonal !== 'false';
