+++ svelte.config.js
@@ -5,7 +5,13 @@ const config = {
     preprocess: vitePreprocess(),
 
     kit: {
-        adapter: adapter()
+        adapter: adapter(),
+        csp: {
+            directives: {
+                'script-src': ["'self'", 'https://apis.google.com', 'https://www.gstatic.com', "'unsafe-inline'", "'unsafe-eval'"],
+                'frame-src': ["'self'", 'https://*.firebaseapp.com', 'https://apis.google.com']
+            }
+        }
     },
 
     vite: {
